+++ src/pmpy/__init__.py
@@ -2,4 +2,4 @@
 """
 print("Hello from pmpy, project management package developed at KNTU (https://github.com/Project-PmPy/pmpy)")
 print("Current subpackage includes des (discrete event simulatoin), dists(distributions)  and mcs (monte carlo simulation) modules.")
-__version__ = "1.0.13"
+__version__ = "1.0.14"